+++ examples/next-ts/pages/date-picker-min-max.tsx
@@ -1,18 +1,16 @@
 import * as datePicker from "@zag-js/date-picker"
 import { normalizeProps, useMachine } from "@zag-js/react"
-import { datePickerControls } from "@zag-js/shared"
 import { useId } from "react"
 import { StateVisualizer } from "../components/state-visualizer"
 import { Toolbar } from "../components/toolbar"
-import { useControls } from "../hooks/use-controls"
 
 export default function Page() {
-  const controls = useControls(datePickerControls)
   const service = useMachine(datePicker.machine, {
     id: useId(),
     locale: "en",
     selectionMode: "single",
-    ...controls.context,
+    min: datePicker.parse("2025-07-01"),
+    max: datePicker.parse("2025-09-30"),
   })
 
   const api = datePicker.connect(service, normalizeProps)
@@ -20,9 +18,6 @@ export default function Page() {
   return (
     <>
       <main className="date-picker">
-        <div>
-          <button>Outside Element</button>
-        </div>
         <p>{`Visible range: ${api.visibleRangeText.formatted}`}</p>
 
         <output className="date-output">
@@ -86,59 +81,11 @@ export default function Page() {
                 </tbody>
               </table>
             </div>
-
-            <div style={{ display: "flex", gap: "40px" }}>
-              <div hidden={api.view !== "month"} style={{ width: "100%" }}>
-                <div {...api.getViewControlProps({ view: "month" })}>
-                  <button {...api.getPrevTriggerProps({ view: "month" })}>Prev</button>
-                  <button {...api.getViewTriggerProps({ view: "month" })}>{api.visibleRange.start.year}</button>
-                  <button {...api.getNextTriggerProps({ view: "month" })}>Next</button>
-                </div>
-
-                <table {...api.getTableProps({ view: "month", columns: 4 })}>
-                  <tbody {...api.getTableBodyProps({ view: "month" })}>
-                    {api.getMonthsGrid({ columns: 4, format: "short" }).map((months, row) => (
-                      <tr key={row} {...api.getTableRowProps()}>
-                        {months.map((month, index) => (
-                          <td key={index} {...api.getMonthTableCellProps({ ...month, columns: 4 })}>
-                            <div {...api.getMonthTableCellTriggerProps({ ...month, columns: 4 })}>{month.label}</div>
-                          </td>
-                        ))}
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-
-              <div hidden={api.view !== "year"} style={{ width: "100%" }}>
-                <div {...api.getViewControlProps({ view: "year" })}>
-                  <button {...api.getPrevTriggerProps({ view: "year" })}>Prev</button>
-                  <span>
-                    {api.getDecade().start} - {api.getDecade().end}
-                  </span>
-                  <button {...api.getNextTriggerProps({ view: "year" })}>Next</button>
-                </div>
-
-                <table {...api.getTableProps({ view: "year", columns: 4 })}>
-                  <tbody {...api.getTableBodyProps()}>
-                    {api.getYearsGrid({ columns: 4 }).map((years, row) => (
-                      <tr key={row} {...api.getTableRowProps({ view: "year" })}>
-                        {years.map((year, index) => (
-                          <td key={index} {...api.getYearTableCellProps({ ...year, columns: 4 })}>
-                            <div {...api.getYearTableCellTriggerProps({ ...year, columns: 4 })}>{year.label}</div>
-                          </td>
-                        ))}
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-            </div>
           </div>
         </div>
       </main>
 
-      <Toolbar viz controls={controls.ui}>
+      <Toolbar viz>
         <StateVisualizer state={service} omit={["weeks"]} />
       </Toolbar>
     </>