+++ examples/next-ts/pages/date-picker-range.tsx
@@ -47,7 +47,7 @@ export default function Page() {
             <div style={{ marginBottom: "20px" }}>
               <select {...api.getMonthSelectProps()}>
                 {api.getMonths().map((month, i) => (
-                  <option key={i} value={i + 1}>
+                  <option key={i} value={i + 1} disabled={month.disabled}>
                     {month.label}
                   </option>
                 ))}
@@ -55,7 +55,7 @@ export default function Page() {
 
               <select {...api.getYearSelectProps()}>
                 {api.getYears().map((year, i) => (
-                  <option key={i} value={year.value}>
+                  <option key={i} value={year.value} disabled={year.disabled}>
                     {year.label}
                   </option>
                 ))}