+++ packages/machines/combobox/src/combobox.types.ts
@@ -55,6 +55,7 @@ export interface OpenChangeDetails {
 export interface ScrollToIndexDetails {
   index: number
   immediate?: boolean | undefined
+  getElement: () => HTMLElement | null
 }
 
 export interface NavigateDetails {