+++ packages/machines/listbox/src/listbox.connect.ts
@@ -7,7 +7,6 @@ import {
   getEventKey,
   getEventTarget,
   getNativeEvent,
-  getWindow,
   isComposingEvent,
   isCtrlOrMetaKey,
   isEditableElement,
@@ -23,16 +22,20 @@ export function connect<T extends PropTypes, V extends CollectionItem = Collecti
   service: Service<ListboxSchema<V>>,
   normalize: NormalizeProps<T>,
 ): ListboxApi<T, V> {
-  const { context, prop, scope, computed, send } = service
+  const { context, prop, scope, computed, send, refs } = service
 
   const disabled = prop("disabled")
   const collection = prop("collection")
   const layout = isGridCollection(collection) ? "grid" : "list"
 
+  const focused = context.get("focused")
+  const focusVisible = refs.get("focusVisible") && focused
+
   const value = context.get("value")
+  const selectedItems = context.get("selectedItems")
+
   const highlightedValue = context.get("highlightedValue")
   const highlightedItem = context.get("highlightedItem")
-  const selectedItems = context.get("selectedItems")
 
   const isTypingAhead = computed("isTypingAhead")
   const interactive = computed("isInteractive")
@@ -43,10 +46,14 @@ export function connect<T extends PropTypes, V extends CollectionItem = Collecti
     const itemDisabled = collection.getItemDisabled(props.item)
     const value = collection.getItemValue(props.item)
     ensure(value, () => `[zag-js] No value found for item ${JSON.stringify(props.item)}`)
+    const highlighted = highlightedValue === value
     return {
       value,
       disabled: Boolean(disabled || itemDisabled),
-      highlighted: highlightedValue === value && context.get("focused"),
+      focused: highlighted && focused,
+      focusVisible: highlighted && focusVisible,
+      // deprecated
+      highlighted: highlighted && focusVisible,
       selected: context.get("value").includes(value),
     }
   }
@@ -115,18 +122,11 @@ export function connect<T extends PropTypes, V extends CollectionItem = Collecti
         enterKeyHint: "go",
         onFocus() {
           queueMicrotask(() => {
-            const contentEl = dom.getContentEl(scope)
-            const win = getWindow(contentEl)
-            const focusInEvt = new win.FocusEvent("focusin", { bubbles: true, cancelable: true })
-            contentEl?.dispatchEvent(focusInEvt)
+            send({ type: "INPUT.FOCUS" })
           })
         },
-        onBlur(event) {
-          if (event.defaultPrevented) return
-          const contentEl = dom.getContentEl(scope)
-          const win = getWindow(contentEl)
-          const focusOutEvt = new win.FocusEvent("focusout", { bubbles: true, cancelable: true })
-          contentEl?.dispatchEvent(focusOutEvt)
+        onBlur() {
+          send({ type: "CONTENT.BLUR", src: "input" })
         },
         onInput(event) {
           if (!props.autoHighlight) return
@@ -345,7 +345,6 @@ export function connect<T extends PropTypes, V extends CollectionItem = Collecti
               }
 
               if (!nextValue) return
-
               event.preventDefault()
               send({ type: "NAVIGATE", value: nextValue, shiftKey, anchorValue: highlightedValue })
             },