+++ packages/machines/scroll-area/src/scroll-area.connect.ts
@@ -1,4 +1,4 @@
-import { dataAttr, getEventPoint } from "@zag-js/dom-query"
+import { contains, dataAttr, getEventPoint, getEventTarget } from "@zag-js/dom-query"
 import type { NormalizeProps, PropTypes } from "@zag-js/types"
 import { toPx } from "@zag-js/utils"
 import { parts } from "./scroll-area.anatomy"
@@ -58,21 +58,26 @@ export function connect<T extends PropTypes>(
         "data-overflow-x": dataAttr(!hiddenState.scrollbarXHidden),
         "data-overflow-y": dataAttr(!hiddenState.scrollbarYHidden),
         onPointerEnter(event) {
+          const target = getEventTarget(event)
+          if (!contains(event.currentTarget, target)) return
           send({ type: "root.pointerenter", pointerType: event.pointerType })
         },
         onPointerMove(event) {
+          const target = getEventTarget(event)
+          if (!contains(event.currentTarget, target)) return
           send({ type: "root.pointerenter", pointerType: event.pointerType })
         },
         onPointerDown({ pointerType }) {
           send({ type: "root.pointerdown", pointerType })
         },
-        onPointerLeave() {
+        onPointerLeave(event) {
+          if (contains(event.currentTarget, event.relatedTarget)) return
           send({ type: "root.pointerleave" })
         },
         style: {
           position: "relative",
-          "--corder-width": toPx(cornerSize?.width),
-          "--corder-height": toPx(cornerSize?.height),
+          "--corner-width": toPx(cornerSize?.width),
+          "--corner-height": toPx(cornerSize?.height),
           "--thumb-width": toPx(thumbSize?.width),
           "--thumb-height": toPx(thumbSize?.height),
         },
@@ -154,12 +159,12 @@ export function connect<T extends PropTypes>(
           userSelect: "none",
           ...(orientation === "vertical" && {
             top: 0,
-            bottom: `var(--corder-height)`,
+            bottom: `var(--corner-height)`,
             insetInlineEnd: 0,
           }),
           ...(orientation === "horizontal" && {
             insetInlineStart: 0,
-            insetInlineEnd: `var(--corder-width)`,
+            insetInlineEnd: `var(--corner-width)`,
             bottom: 0,
           }),
         },
@@ -202,8 +207,8 @@ export function connect<T extends PropTypes>(
           position: "absolute",
           bottom: 0,
           insetInlineEnd: 0,
-          width: "var(--corder-width)",
-          height: "var(--corder-height)",
+          width: "var(--corner-width)",
+          height: "var(--corner-height)",
         },
       })
     },