+++ packages/frameworks/preact/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @zag-js/preact
 
+## 1.21.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/core@1.21.8
+  - @zag-js/store@1.21.8
+  - @zag-js/types@1.21.8
+
 ## 1.21.7
 
 ### Patch Changes